Squashed commit to modularize the runtime
