+++ runtime/include/chpl-gpu.h
@@ -22,6 +22,8 @@
 #define _CHPL_GPU_H_
 
 #include <stdbool.h>
+#include "chpl-tasks.h"
+#include "chpl-mem-desc.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -29,9 +31,23 @@ extern "C" {
 
 #ifdef HAS_GPU_LOCALE
 
+#define ENABLE_GPU_DEBUG 0
+
+static inline void CHPL_GPU_DEBUG(const char *str, ...) {
+#if ENABLE_GPU_DEBUG
+  va_list args;
+  va_start(args, str);
+  vfprintf(stdout, str, args);
+  va_end(args);
+  fflush(stdout);
+#endif
+}
+
+static inline bool chpl_gpu_running_on_gpu_locale(void) {
+  return chpl_task_getRequestedSubloc()>=0;
+}
+
 void chpl_gpu_init(void);
-bool chpl_gpu_has_context(void);
-bool chpl_gpu_running_on_gpu_locale(void);
 
 void chpl_gpu_launch_kernel(int ln, int32_t fn,
                             const char* fatbinData, const char* name,
@@ -55,13 +71,14 @@ void* chpl_gpu_mem_memalign(size_t boundary, size_t size,
                             chpl_mem_descInt_t description,
                             int32_t lineno, int32_t filename);
 void chpl_gpu_mem_free(void* memAlloc, int32_t lineno, int32_t filename);
-size_t chpl_gpu_get_alloc_size(void* ptr);
 
 void chpl_gpu_copy_device_to_host(void* dst, void* src, size_t n);
 void chpl_gpu_copy_host_to_device(void* dst, void* src, size_t n);
 
 bool chpl_gpu_is_device_ptr(void* ptr);
 
+// TODO do we really need to expose this?
+size_t chpl_gpu_get_alloc_size(void* ptr);
 #endif // HAS_GPU_LOCALE
 
 #ifdef __cplusplus
