+++ test/gpu/native/promotion.chpl
@@ -1,3 +1,6 @@
+use GPUDiagnostics;
+
+startGPUDiagnostics();
 config const n = 10;
 on here.gpus[0] {
   var A: [1..n] int;
@@ -6,3 +9,6 @@ on here.gpus[0] {
 
   writeln(A);
 }
+stopGPUDiagnostics();
+
+assert(getGPUDiagnostics()[0].kernel_launch == 1);
