--- test/gpu/native/promotion.execopts
@@ -1,1 +1,0 @@
---verbose
