+++ test/gpu/native/promotion.good
@@ -1,1 +1,2 @@
-1
+warning: The prototype GPU support implies --no-checks. This may impact debuggability. To suppress this warning, compile with --no-checks explicitly
+1 1 1 1 1 1 1 1 1 1
