--- test/gpu/native/promotion.prediff
@@ -1,1 +1,0 @@
-launchCounter.prediff
