--- test/gpu/native/promotion2.execopts
@@ -1,1 +1,0 @@
---verbose
