+++ test/gpu/native/promotion2.good
@@ -1,3 +1,4 @@
+warning: The prototype GPU support implies --no-checks. This may impact debuggability. To suppress this warning, compile with --no-checks explicitly
 Array: 1 1 1 1 1 1 1 1 1 1 
 Array: 2 2 2 2 2 2 2 2 2 2 
 Array: 3 3 3 3 3 3 3 3 3 3 
@@ -5,4 +6,3 @@ Array: 32 32 32 32 32 32 32 32 32 32
 Array: 103 103 103 103 103 103 103 103 103 103 
 Array: 109 109 109 109 109 109 109 109 109 109 
 Array: 110 110 110 110 110 110 110 110 110 110 
-Number of kernel launches: 7
