+++ test/gpu/native/varInInnerBlock.chpl
@@ -1,4 +1,8 @@
+use GPUDiagnostics;
+
 config const n = 10;
+
+startGPUDiagnostics();
 on here.getChild(1) {
   var A: [1..n] int;
 
@@ -14,3 +18,6 @@ on here.getChild(1) {
 
   writeln(A);
 }
+stopGPUDiagnostics();
+
+assert(getGPUDiagnostics()[0].kernel_launch == 1);
