--- test/gpu/native/varInInnerBlock.execopts
@@ -1,1 +1,0 @@
---verbose
