+++ test/gpu/native/varInInnerBlock.good
@@ -1,1 +1,2 @@
-3
+warning: The prototype GPU support implies --no-checks. This may impact debuggability. To suppress this warning, compile with --no-checks explicitly
+110 110 110 110 110 110 110 110 110 110
