+++ util/chplenv/printchplenv.py
@@ -89,8 +89,10 @@ CHPL_ENVS = [
     ChapelEnv('CHPL_TARGET_BACKEND_CPU', INTERNAL),
     ChapelEnv('CHPL_LOCALE_MODEL', RUNTIME | LAUNCHER | DEFAULT, 'loc'),
     ChapelEnv('  CHPL_GPU_CODEGEN', RUNTIME | NOPATH),
+    ChapelEnv('  CHPL_GPU', RUNTIME | DEFAULT, 'gpu'),
     ChapelEnv('  CHPL_CUDA_PATH', RUNTIME | NOPATH),
     ChapelEnv('  CHPL_CUDA_LIBDEVICE_PATH', INTERNAL | NOPATH),
+    ChapelEnv('  CHPL_LIBOMPTARGET_RTL', INTERNAL | NOPATH),
     ChapelEnv('CHPL_COMM', RUNTIME | LAUNCHER | DEFAULT, 'comm'),
     ChapelEnv('  CHPL_COMM_SUBSTRATE', RUNTIME | LAUNCHER | DEFAULT),
     ChapelEnv('  CHPL_GASNET_SEGMENT', RUNTIME | LAUNCHER | DEFAULT),
@@ -179,8 +181,10 @@ def compute_all_values():
 
     ENV_VALS['CHPL_LOCALE_MODEL'] = chpl_locale_model.get()
     ENV_VALS['  CHPL_GPU_CODEGEN'] = chpl_gpu.get()
+    ENV_VALS['  CHPL_GPU'] = chpl_gpu.get_runtime()
     ENV_VALS['  CHPL_CUDA_PATH'] = chpl_gpu.get_cuda_path()
     ENV_VALS['  CHPL_CUDA_LIBDEVICE_PATH'] = chpl_gpu.get_cuda_libdevice_path()
+    ENV_VALS['  CHPL_LIBOMPTARGET_RTL'] = chpl_gpu.get_libomptarget_rtl()
     ENV_VALS['CHPL_COMM'] = chpl_comm.get()
     ENV_VALS['  CHPL_COMM_SUBSTRATE'] = chpl_comm_substrate.get()
     ENV_VALS['  CHPL_GASNET_SEGMENT'] = chpl_comm_segment.get()
